extract generic attribute backed storage models
